Polilinea

+++ working/figura.c
@@ -13,6 +13,9 @@ struct figura{
     figura_t **figura;
     size_t cantidad;
 };*/
+
+
+
 const char* figura_tipo_a_cadena(figura_tipo_t figura){
     static const char* tipo[]={
     [ICONO]="Icono",
@@ -72,13 +75,13 @@ polilinea_t *leer_polilinea(FILE *f){
     return polilinea;
 }
 
-void figura_trasladar(figura_t *figura, double traslacion[2]){
+void figura_trasladar(figura_t *figura, float traslacion[2]){
     for(size_t i=0; i<(figura->cantidad_polilineas);i++){
         polilinea_trasladar(*((figura->polilineas)+i), traslacion);
     }
 }
 
-void figura_rotar(figura_t *figura, double angulo){
+void figura_rotar(figura_t *figura, float angulo){
     for(size_t i=0; i<(figura->cantidad_polilineas);i++){
         polilinea_rotar(*((figura->polilineas)+i), angulo);
     }
