+++ working/polilinea.h
@@ -25,8 +25,8 @@ polilinea_t *get_polilinea(polilinea_t **polilineas,size_t numero_poli,size_t ca
 //Setters
 bool polilinea_setear_punto(polilinea_t *polilinea, size_t pos, float x, float y);
 bool polilinea_setear_color(polilinea_t *polilinea, color_t color);
-void polilinea_trasladar(polilinea_t *polilinea, double traslacion[2]);
-void polilinea_rotar(polilinea_t *polilinea, double angulo);
+void polilinea_trasladar(polilinea_t *polilinea, float traslacion[2]);
+void polilinea_rotar(polilinea_t *polilinea, float angulo);
 bool dibujar_polilinea(SDL_Renderer *renderer,polilinea_t *polilinea, float posicion[2], float escala);
 
 //Memoria
